% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Dmax.R
\name{Dmax}
\alias{Dmax}
\title{Calculate the workload corresponding to the point with the maximum perpendicular
from a line joining the first and last lactate measurments to the estimated lactate
curve.}
\usage{
Dmax(data, workload.column = 1, lactate.column = 2, plot = T)
}
\arguments{
\item{data}{A data frame with work intensity (e.g. watt) and lactate measurements}

\item{workload.column}{An integer specifying the column with work intensity data, default=1.}

\item{lactate.column}{An integer specifying the column with lactate data, default=2.}

\item{plot}{If true, the function returns a plot}
}
\value{
Work intensity at
}
\description{
Calculates the workload at the maximum of the second derivate of the
lactate curve. Based on functions described and written by Newell et al. 2007 (see reference)
}
\references{
Software for calculating blood lactate endurance markers.
John Newell, David Higgins, Niall Madden, James Cruickshank,
Jochen Einbeck, Kenny McMillan, Roddy McDonald.
Journal of Sports Sciences, Vol. 25, Iss. 12, 2007
}

