% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixedLactate.R
\name{fixedLactate}
\alias{fixedLactate}
\title{Calculate and plot lactate threshold based on a fixed lactate concentration}
\usage{
fixedLactate(data, workload.column = 1, lactate.column = 2, plot = T,
  lactate = 4)
}
\arguments{
\item{data}{A data frame with work intensity (e.g. watt) and lactate measurements}

\item{workload.column}{A integer specifying the column with work intensity data, default=1.}

\item{lactate.column}{A integer specifying the column with lactate data, default=2.}

\item{plot}{Logical specify whether to add plot or not. The plot shows data points with the model fit}

\item{lactate}{A number specifying the lacate level at which you want to now the work intensity.}
}
\description{
Calculate and plot lactate threshold based on a fixed lactate concentration
}

